Use unique namespace for each test in garden test
