+++ test/integration/operator/garden/garden_suite_test.go
@@ -18,36 +18,19 @@ import (
 	"context"
 	"path/filepath"
 	"testing"
-	"time"
 
-	druidv1alpha1 "github.com/gardener/etcd-druid/api/v1alpha1"
 	"github.com/go-logr/logr"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/client-go/rest"
-	"k8s.io/utils/pointer"
-	"sigs.k8s.io/controller-runtime/pkg/cache"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/client/apiutil"
 	"sigs.k8s.io/controller-runtime/pkg/envtest"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
-	"sigs.k8s.io/controller-runtime/pkg/manager"
 
-	"github.com/gardener/gardener/charts"
-	gardencorev1beta1 "github.com/gardener/gardener/pkg/apis/core/v1beta1"
-	operatorv1alpha1 "github.com/gardener/gardener/pkg/apis/operator/v1alpha1"
 	"github.com/gardener/gardener/pkg/logger"
-	"github.com/gardener/gardener/pkg/operator/apis/config"
 	operatorclient "github.com/gardener/gardener/pkg/operator/client"
-	"github.com/gardener/gardener/pkg/operator/controller/garden"
 	"github.com/gardener/gardener/pkg/operator/features"
-	"github.com/gardener/gardener/pkg/utils/imagevector"
-	. "github.com/gardener/gardener/pkg/utils/test/matchers"
-	"github.com/gardener/gardener/test/utils/operationannotation"
 )
 
 func TestGarden(t *testing.T) {
@@ -65,9 +48,6 @@ var (
 	testEnv    *envtest.Environment
 	testClient client.Client
 	mgrClient  client.Client
-
-	testRunID     string
-	testNamespace *corev1.Namespace
 )
 
 var _ = BeforeSuite(func() {
@@ -102,74 +82,4 @@ var _ = BeforeSuite(func() {
 	By("Create test client")
 	testClient, err = client.New(restConfig, client.Options{Scheme: operatorclient.RuntimeScheme})
 	Expect(err).NotTo(HaveOccurred())
-
-	By("Create test Namespace")
-	testNamespace = &corev1.Namespace{
-		ObjectMeta: metav1.ObjectMeta{
-			GenerateName: "garden-",
-		},
-	}
-	Expect(testClient.Create(ctx, testNamespace)).To(Succeed())
-	log.Info("Created Namespace for test", "namespaceName", testNamespace.Name)
-	testRunID = testNamespace.Name
-
-	DeferCleanup(func() {
-		By("Delete test Namespace")
-		Expect(testClient.Delete(ctx, testNamespace)).To(Or(Succeed(), BeNotFoundError()))
-	})
-
-	By("Setup manager")
-	mapper, err := apiutil.NewDynamicRESTMapper(restConfig)
-	Expect(err).NotTo(HaveOccurred())
-
-	mgr, err := manager.New(restConfig, manager.Options{
-		Scheme:             operatorclient.RuntimeScheme,
-		MetricsBindAddress: "0",
-		NewCache: cache.BuilderWithOptions(cache.Options{
-			Mapper: mapper,
-			SelectorsByObject: map[client.Object]cache.ObjectSelector{
-				&operatorv1alpha1.Garden{}: {
-					Label: labels.SelectorFromSet(labels.Set{testID: testRunID}),
-				},
-			},
-		}),
-	})
-	Expect(err).NotTo(HaveOccurred())
-	mgrClient = mgr.GetClient()
-
-	// The controller waits for the operation annotation to be removed from Etcd resources, so we need to add a
-	// reconciler for it since envtest does not run the responsible controller (etcd-druid).
-	Expect((&operationannotation.Reconciler{ForObject: func() client.Object { return &druidv1alpha1.Etcd{} }}).AddToManager(mgr)).To(Succeed())
-
-	By("Register controller")
-	chartsPath := filepath.Join("..", "..", "..", "..", charts.Path)
-	imageVector, err := imagevector.ReadGlobalImageVectorWithEnvOverride(filepath.Join(chartsPath, "images.yaml"))
-	Expect(err).NotTo(HaveOccurred())
-
-	Expect((&garden.Reconciler{
-		Config: config.OperatorConfiguration{
-			Controllers: config.ControllerConfiguration{
-				Garden: config.GardenControllerConfig{
-					ConcurrentSyncs: pointer.Int(5),
-					SyncPeriod:      &metav1.Duration{Duration: time.Minute},
-				},
-			},
-		},
-		ImageVector:     imageVector,
-		Identity:        &gardencorev1beta1.Gardener{Name: "test-gardener"},
-		GardenNamespace: testNamespace.Name,
-	}).AddToManager(mgr)).To(Succeed())
-
-	By("Start manager")
-	mgrContext, mgrCancel := context.WithCancel(ctx)
-
-	go func() {
-		defer GinkgoRecover()
-		Expect(mgr.Start(mgrContext)).To(Succeed())
-	}()
-
-	DeferCleanup(func() {
-		By("Stop manager")
-		mgrCancel()
-	})
 })
